// src/app/zdslogic-ui-plugins/contacts/core/interfaces/contactForCreation.model.ts
export interface ContactForCreation {

//  id: string;
//  userId: string;
//  userName: string;
//  ownerId: string;
//	createdAt: string;
//	updatedAt: string;
	fullName: string;	
    firstName: string;
    lastName: string;
	gender: string;
	birthdate: string;
    title: string;
    company: string;
    imageURL: string;
	linkedin: string;
	facebook: string;
    skype: string;
    twitter: string;
    notes: string;
    presenceStatus: string;
    presenceImageUrl: string;
	email: string;
	bio: string
//    enabled: string;
}